server/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
alembic==1.13.3

# AI/ML dependencies
transformers==4.46.0
sentence-transformers==3.2.0
torch>=2.6.0
chromadb==0.5.15
google-generativeai==0.8.3

# ONNX Runtime for optimized inference
optimum[onnxruntime]==1.23.3
onnxruntime==1.20.1

# Authentication
firebase-admin==6.5.0

# Utilities
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
email-validator==2.1.1
pandas==2.2.3

# Testing and development
colorama==0.4.6
